#1 Converted the project from JavaScript to TypeScript.
