+++ .gitignore
@@ -1,3 +1,5 @@
-.idea
-tests
+/.idea
+/.vscode
+/tests
 node_modules
+package-lock.json
