+++ tsconfig.json
@@ -1,0 +1,32 @@
+{
+	"compilerOptions": {
+		"lib"                     : [
+			"ESNext",
+			"dom"
+		],
+		"outDir"                  : "lib",
+		"removeComments"          : false,
+		"target"                  : "ES6",
+		"baseUrl"                 : "./",
+		"esModuleInterop"         : true,
+		"moduleResolution"        : "node",
+		"paths"                   : {},
+		"sourceMap"               : true,
+		"sourceRoot"              : "/",
+		"alwaysStrict"            : true,
+		"allowUnreachableCode"    : false,
+		"noImplicitAny"           : true,
+		"strictNullChecks"        : true,
+		"noImplicitReturns"       : true,
+		"noUncheckedIndexedAccess": true,
+		"noUnusedLocals"          : true,
+		"noUnusedParameters"      : true,
+		"useDefineForClassFields" : true
+	},
+	"include"        : [
+		"./**/*.ts"
+	],
+	"exclude"        : [
+		"node_modules/**/*"
+	]
+}
